FEAT: TopBar 화면 전환 애니메이션 추가
